requirements.txt
pandas
numpy
openpyxl
PyPDF2
faiss-cpu
ollama
sentence-transformers